Made Boid and Obstacle extend Polygon
