+++ src/main/java/boids/objects/Boid.java
@@ -1,5 +1,6 @@
-package boids.drawables;
+package boids.objects;
 
+import boids.drawables.Drawable;
 import boids.drawables.geometry.DPolygon;
 import boids.gui.Animation;
 import boids.math.Rectangle;
@@ -8,61 +9,60 @@ import boids.math.Vector2;
 import java.awt.*;
 import java.util.ArrayList;
 
-public abstract class Boid implements Drawable{
+public abstract class Boid extends DPolygon{
     protected FOV fov;
-    protected Vector2 position;
     protected Vector2 velocity;
     protected Vector2 acceleration;
-    protected final DPolygon shape;
-    protected float size;
     protected int index;
 
     public Boid(float size, Vector2 position, FOV fov, Color color) {
-        this.size = size;
-        this.position = position;
-        this.fov = fov;
-        velocity = Vector2.ZERO;
-        shape = new DPolygon(
-                new Vector2[]{
+        super(new Vector2[]{
                         new Vector2(0f, 0f),
                         new Vector2(0.5f * size, 1f * size),
                         new Vector2(0f, 0.75f * size),
                         new Vector2(-0.5f * size, 1f * size)},
                 position, color);
+        this.position = position;
+        this.fov = fov;
+        velocity = Vector2.ZERO;
     }
 
     private void rotate() {
         if (velocity.magnitude() == 0) {
             return;
         }
-        Vector2 direction = position.subtract(shape.getVertices()[2]);
+        Vector2 direction = position.subtract(getVertices()[2]);
         Vector2 newDirection = velocity;
 
         if (direction.isZero() || newDirection.isZero()) {
             return;
         }
         float angle = direction.directAngle(newDirection);
-        shape.rotate(angle);
+        rotate(angle);
         fov.rotate(angle);
     }
 
     private void remainOnScreen(Rectangle frame) {
-        if (position.x > frame.max.x) {
-            position.x = frame.min.x;
-        } else if (position.x < frame.min.x) {
-            position.x = frame.max.x;
+        Vector2 newPosition = new Vector2(position);
+        if (newPosition.x > frame.max.x) {
+            newPosition.x = frame.min.x;
+        } else if (newPosition.x < frame.min.x) {
+            newPosition.x = frame.max.x;
         }
-        if (position.y > frame.max.y) {
-            position.y = frame.min.y;
-        } else if (position.y < frame.min.y) {
-            position.y = frame.max.y;
+        if (newPosition.y > frame.max.y) {
+            newPosition.y = frame.min.y;
+        } else if (newPosition.y < frame.min.y) {
+            newPosition.y = frame.max.y;
+        }
+        if (!newPosition.equals(position)) {
+            moveTo(newPosition);
         }
     }
 
     private Obstacle findWorst(ArrayList<Drawable> objects) {
         float minDistance = Float.MAX_VALUE;
         Obstacle minObstacle = null;
-        for (var i : Obstacle.obstacles) {
+        for (var i : Obstacle.obstaclesIndices) {
             Obstacle obstacle = (Obstacle) objects.get(i);
             if (fov.isIntersecting(obstacle.getBoundingBox(), velocity)) {
                 float distance;
@@ -97,19 +97,23 @@ public abstract class Boid implements Drawable{
         return steer;
     }
 
+    public Vector2 getPosition() {
+        return new Vector2(position);
+    }
+
     @Override
     public void update(Animation animation, double frameTime) {
-        position = position.add(velocity.multiply(frameTime));
-        remainOnScreen(animation.getDimensions());
-        shape.moveTo(position);
+        Vector2 newPosition = position.add(velocity.multiply(frameTime));
+        moveTo(newPosition);
         rotate();
-        fov.setPosition(position);
+        remainOnScreen(animation.getDimensions());
+        fov.setPosition(newPosition);
         fov.setDirection(velocity);
     }
 
     @Override
     public void render(Graphics2D g2d) {
-        shape.render(g2d);
+        super.render(g2d);
     }
 
     protected void setIndex(int index) {
