+++ src/main/java/boids/objects/Prey.java
@@ -1,18 +1,17 @@
-package boids.drawables;
+package boids.objects;
 
 import boids.Utils;
+import boids.drawables.Drawable;
 import boids.gui.Animation;
 import boids.gui.AnimationPanel;
 import boids.math.Vector2;
 
 import java.awt.*;
 import java.util.ArrayList;
-import java.util.Random;
-import java.util.concurrent.ThreadLocalRandom;
 
 public class Prey extends Boid {
 
-    protected static final ArrayList<Integer> preys = new ArrayList<>();
+    protected static final ArrayList<Integer> preysIndices = new ArrayList<>();
     private static final float maxSpeed = 50f;
     private static final float maxAcceleration = 10f;
     private static final float desiredSeparation = 30f;
@@ -34,7 +33,7 @@ public class Prey extends Boid {
         Vector2 steer = Vector2.ZERO;
         int count = 0;
 
-        for (var i : preys) {
+        for (var i : preysIndices) {
             if (i == index) {
                 continue;
             }
@@ -65,7 +64,7 @@ public class Prey extends Boid {
         Vector2 steer = Vector2.ZERO;
         int count = 0;
 
-        for (var i : preys) {
+        for (var i : preysIndices) {
             if (i == index) {
                 continue;
             }
@@ -93,7 +92,7 @@ public class Prey extends Boid {
         Vector2 steer = Vector2.ZERO;
         int count = 0;
 
-        for (var i : preys) {
+        for (var i : preysIndices) {
             if (i == index) {
                 continue;
             }
@@ -120,14 +119,14 @@ public class Prey extends Boid {
         Vector2 steer = Vector2.ZERO;
         int count = 0;
 
-        for (var i : Predator.predators) {
+        for (var i : Predator.predatorsIndices) {
             if (i == index) {
                 continue;
             }
             Predator predator = (Predator) objects.get(i);
-            if (fov.isIntersecting(predator.position)) {
-                float distance = Vector2.distance(this.position, predator.position);
-                Vector2 diff = this.position.subtract(predator.position);
+            if (fov.isIntersecting(predator.getPosition())) {
+                float distance = Vector2.distance(position, predator.getPosition());
+                Vector2 diff = this.position.subtract(predator.getPosition());
                 if (distance * distance > 0) {
                     diff = diff.divide(distance * distance);
                 }
@@ -156,7 +155,7 @@ public class Prey extends Boid {
                     Utils.randomFloat(0f, panel.getWidth()),
                     Utils.randomFloat(0f, panel.getHeight())
             ));
-            for (var i : Obstacle.obstacles) {
+            for (var i : Obstacle.obstaclesIndices) {
                 if (prey.position.isInside(((Obstacle)objects.get(i)).getBoundingBox())) {
                     intersects = true;
                 }
@@ -169,20 +168,20 @@ public class Prey extends Boid {
         );
 
         objects.add(prey);
-        preys.add(objects.size() - 1);
+        preysIndices.add(objects.size() - 1);
         prey.setIndex(objects.size() - 1);
     }
 
     public static void removePrey(ArrayList<Drawable> objects) {
-        if (preys.size() > 0) {
-            objects.remove((int) preys.get(preys.size() - 1));
-            preys.remove(preys.size() - 1);
+        if (preysIndices.size() > 0) {
+            objects.remove((int) preysIndices.get(preysIndices.size() - 1));
+            preysIndices.remove(preysIndices.size() - 1);
         }
     }
     public static float getAverageVelocity(ArrayList<Drawable> objects){
         float averageVelocity = 0;
         int count = 0;
-        for (var i : preys) {
+        for (var i : preysIndices) {
             Prey prey = (Prey) objects.get(i);
             averageVelocity += prey.velocity.magnitude();
             count++;
